Isolated ROS node from class definition
