+++ CMakeLists.txt
@@ -157,7 +157,7 @@ include_directories(
 ## in contrast to setup.py, you can choose the destination
 install(PROGRAMS
   scripts/ros_works_turtlesim/turtle_open_control_node
-  scripts/ros_works_turtlesim/turtle_closed_loop
+  scripts/ros_works_turtlesim/turtle_closed_loop_node
   scripts/ros_works_turtlesim_swarm/turtle_swarm_node
   DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
 )
